This enables tracing
Within vscode you can go into the user profile's settings
and add stuff like:
    "json.trace.server": "verbose"

But if we do it for 'sc' then it won't work without this extension configuration
we're adding.

This is for controlling Output panel messages which show sent and received LSP messages,
and it's great for debugging.